requirements/requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-login

python-dotenv